+++ src/mixins/filterMessages.js
@@ -55,7 +55,7 @@ export default {
                 return flag && !!message[filter.field] && message[filter.field] >= filter.value
               }
               case '=': {
-                let val = JSON.stringify(message[filter.field])
+                let val = JSON.stringify(message[filter.field]).slice(1, -1)
                 return flag && !!val && !!val.match(`^${filter.value.replace(/\*/g, '.*')}$`)
               }
               case '<': {