+++ src/pages/viewer/Channels.vue
@@ -84,41 +84,42 @@
             <q-tooltip>View hex payload</q-tooltip>
           </q-icon>
         </div>
-        <q-btn v-if="!selectedItem.deleted" flat class="on-left" color="white" @click="modeModel = !modeModel" :icon="modeModel ? 'playlist_play' : 'history'" :rounded="$q.platform.is.mobile">
-          {{$q.platform.is.mobile ? '' : modeModel ? 'Real-time' : 'History'}}
-          <q-chip small square color="red" v-if="newMessagesCount" class="cursor-pointer q-ml-sm">{{newMessagesCount}}</q-chip>
-          <q-tooltip>Mode (Real-time/History)</q-tooltip>
-        </q-btn>
+        <div>
+          <q-btn v-if="!selectedItem.deleted" flat class="on-left" color="white" @click="modeModel = !modeModel" :icon="modeModel ? 'playlist_play' : 'history'" :rounded="$q.platform.is.mobile">
+            {{$q.platform.is.mobile ? '' : modeModel ? 'Real-time' : 'History'}}
+            <q-chip small square color="red" v-if="newMessagesCount" class="cursor-pointer q-ml-sm">{{newMessagesCount}}</q-chip>
+            <q-tooltip>Mode (Real-time/History)</q-tooltip>
+          </q-btn>
+          <q-btn-toggle
+            v-if="!selectedItem.deleted"
+            dense
+            color="grey-8"
+            toggle-color="white"
+            toggle-text-color="dark"
+            class="q-ml-sm gt-xs" size="sm"
+            v-model="ratio"
+            :options="[{label: 'logs', value: 100},{label: 'both', value: 50},{label: 'messages', value: 0}]"
+          />
+          <q-btn class="lt-sm" dense size="sm">
+            {{ratio === 50 ? 'both' : (ratio === 0 ? 'messages' : 'logs')}}
+            <q-popover style="background-color: transparent">
+              <q-btn-toggle
+                v-close-overlay
+                v-if="!selectedItem.deleted"
+                dense
+                color="grey-8"
+                toggle-color="white"
+                toggle-text-color="dark"
+                size="sm"
+                v-model="ratio"
+                :options="[{label: 'logs', value: 100},{label: 'both', value: 50},{label: 'messages', value: 0}]"
+              />
+            </q-popover>
+          </q-btn>
+        </div>
         <div>
           <q-icon size="1.5rem" class="on-left cursor-pointer pull-right" v-if="modeModel && !isEmptyMessages" color="white" name="mdi-playlist-remove" @click.native="clearHandler">
             <q-tooltip>Clear all panes</q-tooltip>
-          </q-icon>
-          <q-icon v-if="!selectedItem.deleted" size="1.5rem" class="cursor-pointer pull-right" name="mdi-format-align-middle">
-            <q-tooltip>Section ratio</q-tooltip>
-            <q-popover ref="ratioPopover">
-              <q-item style="width: 25rem; height: 100px" class="bg-dark">
-                <q-item-side class="text-center">
-                  <q-item-tile color="grey-6">Logs</q-item-tile>
-                </q-item-side>
-                <q-item-main>
-                  <q-item-tile label class="ellipsis overflow-hidden" color="white">Ratio</q-item-tile>
-                  <q-item-tile sublabel>
-                    <q-slider
-                      v-model="ratio"
-                      color="grey-6"
-                      :min="0"
-                      :max="100"
-                      :step="25"
-                      label
-                      snap
-                    />
-                  </q-item-tile>
-                </q-item-main>
-                <q-item-side class="text-center" right>
-                  <q-item-tile color="grey-6">Messages</q-item-tile>
-                </q-item-side>
-              </q-item>
-            </q-popover>
           </q-icon>
         </div>
       </q-toolbar>